+++ Model/Operation/CsvImportByUrlOperation.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * @author InRiver <iif-magento@inriver.com>
+ * @author InRiver <inriveradapters@inriver.com>
  * @copyright Copyright (c) InRiver (https://www.inriver.com/)
  * @link https://www.inriver.com/
  */
@@ -21,8 +21,10 @@ use Magento\Framework\App\Config\ScopeConfigInterface;
 use Magento\Framework\App\Filesystem\DirectoryList;
 use Magento\Framework\Exception\LocalizedException;
 use Magento\Framework\Filesystem;
+use Magento\Store\Api\WebsiteRepositoryInterface;
 
 use function __;
+use function date;
 use function ltrim;
 
 /**
@@ -48,25 +50,31 @@ class CsvImportByUrlOperation implements ProductsImportInterface
     /** @var \Inriver\Adapter\Helper\FileEncoding */
     protected $fileEncoding;
 
+    /** @var \Magento\Store\Api\WebsiteRepositoryInterface  */
+    private $websiteRepository;
+
     /**
      * @param \Magento\Framework\App\Config\ScopeConfigInterface $scopeConfig
      * @param \Inriver\Adapter\Model\Data\ImportFactory $importFactory
      * @param \Inriver\Adapter\Helper\FileDownloader $downloader
      * @param \Magento\Framework\Filesystem $filesystem
      * @param \Inriver\Adapter\Helper\FileEncoding $fileEncoding
+     * @param \Magento\Store\Api\WebsiteRepositoryInterface $storeManager
      */
     public function __construct(
         ScopeConfigInterface $scopeConfig,
         ImportFactory $importFactory,
         FileDownloader $downloader,
         Filesystem $filesystem,
-        FileEncoding $fileEncoding
+        FileEncoding $fileEncoding,
+        WebsiteRepositoryInterface  $websiteRepository
     ) {
         $this->scopeConfig = $scopeConfig;
         $this->importFactory = $importFactory;
         $this->downloader = $downloader;
         $this->filesystem = $filesystem;
         $this->fileEncoding = $fileEncoding;
+        $this->websiteRepository = $websiteRepository;
     }
 
     /**
@@ -75,6 +83,7 @@ class CsvImportByUrlOperation implements ProductsImportInterface
      * @param \Inriver\Adapter\Api\Data\ProductsImportRequestInterface $import
      *
      * @return \Inriver\Adapter\Api\Data\OperationResultInterface[]
+     * @throws \Magento\Framework\Exception\FileSystemException
      * @throws \Magento\Framework\Exception\LocalizedException
      */
     public function post(ProductsImportRequestInterface $import): array
@@ -85,8 +94,9 @@ class CsvImportByUrlOperation implements ProductsImportInterface
         $output = $this->filesystem->getDirectoryRead(DirectoryList::VAR_DIR);
         $fullPath = $output->getAbsolutePath($filename);
         $this->fileEncoding->removeUtf8Bom($fullPath);
+        $managedWebsiteId = $this->getManagedWebsiteIds($import->getManagedWebsites());
 
-        return $this->startImport($fullPath);
+        return $this->startImport($fullPath, $managedWebsiteId);
     }
 
     /**
@@ -98,7 +108,8 @@ class CsvImportByUrlOperation implements ProductsImportInterface
      */
     protected function getCsvFile(): string
     {
-        $destination = $this->getTargetDirectory() . '/import.csv';
+
+        $destination = $this->getTargetDirectory() . '/import-' . date('Ymdhis') . '.csv';
         $bytesWritten = $this->downloader->download($this->sourceUrl, $destination);
 
         if ($bytesWritten === 0) {
@@ -135,14 +146,33 @@ class CsvImportByUrlOperation implements ProductsImportInterface
 
     /**
      * @param string $path
+     * @param string $managedWebsiteIds
      *
      * @return array
      */
-    protected function startImport(string $path): array
+    protected function startImport(string $path, string $managedWebsiteIds): array
     {
         $import = $this->importFactory->create();
+        $import->setManagedWebsites($managedWebsiteIds);
         $import->execute($path);
 
         return $import->getErrorsAsArray();
+    }
+
+    /**
+     *
+     */
+    private function getManagedWebsiteIds(?array $managedWebsites): string
+    {
+        if($managedWebsites !== null) {
+            $ids = array();
+            foreach ($managedWebsites as $code) {
+                $ids[] = $this->websiteRepository->get($code)->getId();
+            }
+
+            return implode(',', $ids);
+        }
+
+        return  '';
     }
 }