+++ Model/Operation/CsvProductRelationsImportByUrlOperation.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * @author InRiver <iif-magento@inriver.com>
+ * @author InRiver <inriveradapters@inriver.com>
  * @copyright Copyright (c) InRiver (https://www.inriver.com/)
  * @link https://www.inriver.com/
  */
@@ -17,6 +17,7 @@ use Inriver\Adapter\Model\Data\ImportFactory;
 use Inriver\Adapter\Model\Data\ImportProductRelationsFactory;
 use Magento\Framework\App\Config\ScopeConfigInterface;
 use Magento\Framework\Filesystem;
+use Magento\Store\Api\WebsiteRepositoryInterface;
 
 /**
  * Class CsvProductRelationsImportByUrlOperation Csv Product Relations Import by url Operation
@@ -33,6 +34,7 @@ class CsvProductRelationsImportByUrlOperation extends CsvImportByUrlOperation im
      * @param \Magento\Framework\Filesystem $filesystem
      * @param \Inriver\Adapter\Helper\FileEncoding $fileEncoding
      * @param \Inriver\Adapter\Model\Data\ImportProductRelationsFactory $importProductRelationsFactory
+     * @param \Magento\Store\Api\WebsiteRepositoryInterface $websiteRepository
      */
     public function __construct(
         ScopeConfigInterface $scopeConfig,
@@ -40,14 +42,21 @@ class CsvProductRelationsImportByUrlOperation extends CsvImportByUrlOperation im
         FileDownloader $downloader,
         Filesystem $filesystem,
         FileEncoding $fileEncoding,
-        ImportProductRelationsFactory $importProductRelationsFactory
+        ImportProductRelationsFactory $importProductRelationsFactory,
+        WebsiteRepositoryInterface  $websiteRepository
     ) {
         parent::__construct(
             $this->scopeConfig = $scopeConfig,
             $this->importFactory = $importFactory,
             $this->downloader = $downloader,
             $this->filesystem = $filesystem,
-            $this->fileEncoding = $fileEncoding
+            $this->fileEncoding = $fileEncoding,
+            $scopeConfig,
+            $importFactory,
+            $downloader,
+            $filesystem,
+            $fileEncoding,
+            $websiteRepository
         );
 
         $this->importProductRelationsFactory = $importProductRelationsFactory;