+++ Model/Operation/ProductImagesOperation.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * @author InRiver <iif-magento@inriver.com>
+ * @author InRiver <inriveradapters@inriver.com>
  * @copyright Copyright (c) InRiver (https://www.inriver.com/)
  * @link https://www.inriver.com/
  */
@@ -24,6 +24,7 @@ use Magento\Catalog\Api\Data\ProductAttributeMediaGalleryEntryInterface;
 use Magento\Catalog\Api\Data\ProductAttributeMediaGalleryEntryInterfaceFactory;
 use Magento\Catalog\Api\Data\ProductInterface;
 use Magento\Catalog\Model\ProductRepository;
+use Magento\CatalogInventory\Model\StockRegistryStorage;
 use Magento\Framework\Api\Data\ImageContentInterfaceFactory;
 use Magento\Framework\App\Filesystem\DirectoryList;
 use Magento\Framework\Exception\LocalizedException;
@@ -72,6 +73,9 @@ class ProductImagesOperation implements ImagesInterface
     /** @var \Inriver\Adapter\Model\MediaGallery\MediaGalleryManagement */
     protected $mediaGalleryManagement;
 
+    /** @var \Magento\CatalogInventory\Model\StockRegistryStorage */
+    private $stockRegistryStorage;
+
     /** @var \Inriver\Adapter\Logger\Logger */
     private $logger;
 
@@ -85,6 +89,7 @@ class ProductImagesOperation implements ImagesInterface
      * @param \Inriver\Adapter\Api\InriverMediaGalleryDataRepositoryInterface $inriverMediaGalleryDataRepository
      * @param \Inriver\Adapter\Model\Data\InriverMediaGalleryDataFactory $inriverMediaGalleryDataFactory
      * @param \Inriver\Adapter\Model\MediaGallery\MediaGalleryManagement $mediaGalleryManagement
+     * @param \Magento\CatalogInventory\Model\StockRegistryStorage $stockRegistryStorage
      * @param \Inriver\Adapter\Logger\Logger $logger
      */
     public function __construct(
@@ -97,6 +102,7 @@ class ProductImagesOperation implements ImagesInterface
         InriverMediaGalleryDataRepositoryInterface $inriverMediaGalleryDataRepository,
         InriverMediaGalleryDataFactory $inriverMediaGalleryDataFactory,
         MediaGalleryManagement $mediaGalleryManagement,
+        StockRegistryStorage $stockRegistryStorage,
         Logger $logger
     ) {
         $this->downloader = $downloader;
@@ -108,6 +114,7 @@ class ProductImagesOperation implements ImagesInterface
         $this->inriverMediaGalleryDataRepository = $inriverMediaGalleryDataRepository;
         $this->inriverMediaGalleryDataFactory = $inriverMediaGalleryDataFactory;
         $this->mediaGalleryManagement = $mediaGalleryManagement;
+        $this->stockRegistryStorage = $stockRegistryStorage;
         $this->logger = $logger;
     }
 
@@ -143,12 +150,17 @@ class ProductImagesOperation implements ImagesInterface
      * @param \Inriver\Adapter\Api\Data\ProductImagesInterface\ImageInterface[] $images
      *
      * @return string[]
-     * @throws \Magento\Framework\Exception\LocalizedException
      * @throws \Magento\Framework\Exception\NoSuchEntityException
+     * @throws \Magento\Framework\Exception\StateException
+     * @throws \Magento\Framework\Exception\InputException|\Magento\Framework\Exception\LocalizedException
      */
     private function syncProductImages(string $sku, array $images): array
     {
         $this->storeManager->setCurrentStore(0);
+
+        // Clean both stock registry and product repository cache to force a reload.
+        // This Fixes a rare bug if another extension loaded wrong data in them.
+        $this->stockRegistryStorage->clean();
         $this->productRepository->cleanCache();
 
         $newImagesByImageId = $this->imagesByImageId($images);
@@ -221,9 +233,10 @@ class ProductImagesOperation implements ImagesInterface
      * @param \Inriver\Adapter\Api\Data\ProductImagesInterface\ImageInterface $newImage
      * @param string $sku
      *
+     * @throws \Magento\Framework\Exception\FileSystemException
+     * @throws \Magento\Framework\Exception\InputException
      * @throws \Magento\Framework\Exception\LocalizedException
      * @throws \Magento\Framework\Exception\NoSuchEntityException
-     * @throws \Magento\Framework\Exception\InputException
      * @throws \Magento\Framework\Exception\StateException
      */
     private function createNewMediaEntry(ImageInterface $newImage, string $sku): void
@@ -303,8 +316,8 @@ class ProductImagesOperation implements ImagesInterface
      * @param string $sku
      * @param \Inriver\Adapter\Api\Data\ProductImagesInterface\ImageInterface $newImage
      *
-     * @throws \Magento\Framework\Exception\LocalizedException
      * @throws \Magento\Framework\Exception\NoSuchEntityException
+     * @throws \Magento\Framework\Exception\LocalizedException
      */
     private function updateStoreValues(
         ProductAttributeMediaGalleryEntryInterface $entry,
@@ -371,8 +384,8 @@ class ProductImagesOperation implements ImagesInterface
     /**
      * Test if image types are equal
      *
-     * @param array $currentTypes
-     * @param array $requiredTypes
+     * @param string[] $currentTypes
+     * @param string[] $requiredTypes
      *
      * @return bool
      */