+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+##Version 23.4.4
+- Fix code to allow virtual product
+- Fix to product import to keep website associated to a product that is not managed by the adapter
+- Fix the email in the source code
+- Fix code to allow virtual product
+- Fix product import to allow websites not managed by the adapter
+- Added Archiving of imported file for product import
+- Added a new config: "Archive Lifetime in days" (default: 30)
+
 ##Version 23.4.3
 - Fix an error with the implementation of endpoint products/import/relations
 - Fix the list of special characters according to the list in the inRiver adapter